feat: set tools executed in base to be able to track tools executed for a specialist after a handoff from router
