+++ packages/botonic-plugin-ai-agents/src/runners/base-runner.ts
@@ -3,13 +3,18 @@ import type {
   ResolvedPlugins,
   ToolExecution,
 } from '@botonic/core'
-import { InputGuardrailTripwireTriggered, Runner } from '@openai/agents'
-
+import {
+  InputGuardrailTripwireTriggered,
+  Runner,
+  RunToolCallItem,
+  RunToolCallOutputItem,
+} from '@openai/agents'
 import { isProd } from '../constants'
 import type { DebugLogger } from '../debug-logger'
 import { getApiVersion, type LLMConfig } from '../llm-config'
 import { HubtypeApiClient } from '../services/hubtype-api-client'
 import { TrackFeature, TrackProductName } from '../services/types'
+import { RETRIEVE_KNOWLEDGE_TOOL_NAME } from '../tools'
 import type {
   AgenticInputMessage,
   AgenticOutputMessage,
@@ -109,10 +114,105 @@ export abstract class BaseRunner<
   }
 
   protected getToolsExecuted(
-    _result: RunnerResult,
-    _context: Context<TPlugins, TExtraData>
+    result: RunnerResult,
+    context: Context<TPlugins, TExtraData>
   ): ToolExecution[] {
-    return []
+    const toolResultsByCallId = this.getToolResultsByCallId(result.newItems)
+
+    return (
+      result.newItems
+        ?.filter(item => item instanceof RunToolCallItem)
+        .map((item: RunToolCallItem) =>
+          this.getToolExecutionInfo(
+            item as RunToolCallItem,
+            context,
+            toolResultsByCallId
+          )
+        )
+        .filter(
+          (toolExecution: ToolExecution) => toolExecution.toolName !== ''
+        ) || []
+    )
+  }
+
+  private getToolResultsByCallId(
+    newItems: unknown[] = []
+  ): Map<string, string> {
+    const map = new Map<string, string>()
+    for (const item of newItems) {
+      if (!(item instanceof RunToolCallOutputItem)) {
+        continue
+      }
+      const rawItem = item.rawItem as {
+        callId?: string
+        output?: string | { type?: string; text?: string }
+      }
+      const callId = rawItem?.callId
+      const output = rawItem?.output
+      if (callId == null || output == null) {
+        continue
+      }
+      const text = this.extractToolResult(output)
+      if (text != null) {
+        map.set(callId, text)
+      }
+    }
+    return map
+  }
+
+  private extractToolResult(
+    output: string | { type?: string; text?: string }
+  ): string | undefined {
+    if (typeof output === 'string') {
+      return output
+    }
+    return output?.type === 'text' && typeof output?.text === 'string'
+      ? output.text
+      : undefined
+  }
+
+  private getToolExecutionInfo(
+    item: RunToolCallItem,
+    context: Context<TPlugins, TExtraData>,
+    toolResultsByCallId: Map<string, string>
+  ): ToolExecution {
+    if (item.rawItem.type !== 'function_call') {
+      return {
+        toolName: '',
+        toolArguments: {},
+      }
+    }
+    const toolName = item.rawItem.name
+    const toolArguments = this.getSafeToolArguments(item.rawItem.arguments)
+    const toolResults = item.rawItem.callId
+      ? toolResultsByCallId.get(item.rawItem.callId)
+      : undefined
+
+    const toolExecution: ToolExecution = {
+      toolName,
+      toolArguments,
+      toolResults,
+    }
+
+    if (toolName === RETRIEVE_KNOWLEDGE_TOOL_NAME) {
+      return {
+        ...toolExecution,
+        knowledgebaseSourcesIds: context.knowledgeUsed.sourceIds,
+        knowledgebaseChunksIds: context.knowledgeUsed.chunksIds,
+      }
+    }
+
+    return toolExecution
+  }
+
+  private getSafeToolArguments(
+    rawToolArguments: string
+  ): Record<string, unknown> {
+    try {
+      return JSON.parse(rawToolArguments)
+    } catch (_error) {
+      return {}
+    }
   }
 
   protected buildRunResult(
